+++ lib/delay.c
@@ -9,7 +9,7 @@ void Delay10us(unsigned char time)
         while(time--);
 }
 
-void Delay1ms(unsigned int time)
+void Delay1ms(unsigned short time)
 {
         unsigned char i;
         while(time--)
